+++ platform-modules/content-manager/src/main/java/org/ekstep/content/util/ContentPackageExtractionUtil.java
@@ -1,9 +1,13 @@
 package org.ekstep.content.util;
 
+import akka.actor.ActorRef;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.ekstep.async.enums.AsyncActorNames;
+import org.ekstep.async.router.AsyncRequestRouterPool;
 import org.ekstep.common.Platform;
 import org.ekstep.common.Slug;
+import org.ekstep.common.dto.Request;
 import org.ekstep.common.exception.ClientException;
 import org.ekstep.common.exception.ServerException;
 import org.ekstep.common.util.HttpDownloadUtility;
@@ -16,7 +20,9 @@ import org.ekstep.content.enums.ContentErrorCodeConstants;
 import org.ekstep.graph.dac.model.Node;
 import org.ekstep.learning.common.enums.ContentAPIParams;
 import org.ekstep.learning.common.enums.ContentErrorCodes;
-import org.ekstep.learning.util.CloudStore;
+import org.ekstep.learning.util.cloud.CloudStore;
+import org.ekstep.learning.util.cloud.enums.CloudStoreOperations;
+import org.ekstep.learning.util.cloud.enums.CloudStoreParams;
 import org.ekstep.telemetry.logger.TelemetryManager;
 import scala.Option;
 
@@ -187,7 +193,7 @@ public class ContentPackageExtractionUtil {
 				}
 
 				// upload Extracted Content Package
-				uploadExtractedPackage(contentId, node, extractionBasePath, extractionType, slugFile);
+				uploadExtractedPackage(contentId, node, extractionBasePath, extractionType, slugFile, false);
 			} catch (IOException e) {
 				TelemetryManager.error("Error! While unzipping the content package file: "+ e.getMessage(), e);
 			} catch (Exception e) {
@@ -216,7 +222,7 @@ public class ContentPackageExtractionUtil {
 	 *            the extraction type
 	 */
 	public void uploadExtractedPackage(String contentId, Node node, String basePath, ExtractionType extractionType,
-			boolean slugFile) {
+			boolean slugFile, boolean async) {
 
 		if (StringUtils.isBlank(basePath))
 			throw new ClientException(ContentErrorCodes.UPLOAD_DENIED.name(),
@@ -231,17 +237,18 @@ public class ContentPackageExtractionUtil {
 			//Get Cloud folder
 			cloudExtractionPath = getExtractionPath(contentId, node, extractionType);
 			// Upload Directory to Cloud
-			directoryUpload(extractedDir, cloudExtractionPath, slugFile);
+			directoryUpload(extractedDir, cloudExtractionPath, slugFile, basePath, async);
 		}  catch (Exception e) {
 			cleanUpCloudFolder(cloudExtractionPath);
 			throw new ServerException(ContentErrorCodes.EXTRACTION_ERROR.name(),
 					"Error! Something went wrong while extracting the Content Package on Storage Space.", e);
 		} finally {
 			try {
-				TelemetryManager.log("Deleting Locally Extracted File.");
 				File dir = new File(basePath);
-				if (dir.exists())
+				if (!async && dir.exists()) {
+					TelemetryManager.log("Deleting Locally Extracted File | File Path :: " + dir.getAbsolutePath());
 					dir.delete();
+				}
 			} catch (SecurityException e) {
 				TelemetryManager.error("Error! While deleting the local extraction directory: " + basePath, e);
 			} catch (Exception e) {
@@ -275,21 +282,50 @@ public class ContentPackageExtractionUtil {
 	 *            the Cloud folder path
 	 * @return the list
 	 */
-	private void directoryUpload(File directory, String cloudFolderPath, boolean slugFile) {
-		// Validating directory to be uploaded
-		if (!directory.exists())
-			throw new ClientException(ContentErrorCodes.UPLOAD_DENIED.name(),
-					"Error! Extraction File cannot be Empty or 'null' for Content Package Extraction over Storage Space.");
-		if (directory.isFile())
-			throw new ClientException(ContentErrorCodes.UPLOAD_DENIED.name(),
-					"Error! Not a Directory");
-		if (directory.listFiles().length < 1)
-			throw new ClientException(ContentErrorCodes.UPLOAD_DENIED.name(),
-					"Error! Atleast One file is needed for Content Package Extraction.");
+	private void directoryUpload(File directory, String cloudFolderPath, boolean slugFile, String basePath, boolean async) {
+        // Validating directory to be uploaded
+        if (!directory.exists())
+            throw new ClientException(ContentErrorCodes.UPLOAD_DENIED.name(),
+                    "Error! Extraction File cannot be Empty or 'null' for Content Package Extraction over Storage Space.");
+        if (directory.isFile())
+            throw new ClientException(ContentErrorCodes.UPLOAD_DENIED.name(),
+                    "Error! Not a Directory");
+        if (directory.listFiles().length < 1)
+            throw new ClientException(ContentErrorCodes.UPLOAD_DENIED.name(),
+                    "Error! Atleast One file is needed for Content Package Extraction.");
 
-		TelemetryManager.log("Uploading Extracted Directory to Cloud");
-		TelemetryManager.log("Folder Name For Storage Space Extraction: " + cloudFolderPath);
-		CloudStore.uploadDirectory(cloudFolderPath, directory, slugFile);
+        TelemetryManager.log("Uploading Extracted Directory to Cloud");
+        TelemetryManager.log("Folder Name For Storage Space Extraction: " + cloudFolderPath);
+
+        if (async) directoryUploadAsync(directory, cloudFolderPath, slugFile, basePath);
+        else CloudStore.uploadDirectory(cloudFolderPath, directory, slugFile);
+	}
+
+	/**
+	 * Directory upload Asynchronously.
+	 *
+	 * @param directory
+	 *            the directory to be uploaded
+	 * @param cloudFolderPath
+	 *            the Cloud folder path
+	 * @return the list
+	 */
+	private void directoryUploadAsync(File directory, String cloudFolderPath, boolean slugFile, String basePath) {
+		System.out.println("Uploading extracted folder to :: " + cloudFolderPath);
+		Request request = new Request();
+		request.setManagerName(AsyncActorNames.CloudStoreAsync.name());
+		request.setOperation(CloudStoreOperations.uploadDirectory.name());
+		request.put(CloudStoreParams.file.name(), directory);
+		request.put(CloudStoreParams.folderName.name(), cloudFolderPath);
+		request.put(CloudStoreParams.slugFile.name(), slugFile);
+		request.put(CloudStoreParams.basePath.name(), basePath);
+		request.put(CloudStoreParams.deleteFile.name(), true);
+		ActorRef router = AsyncRequestRouterPool.getRequestRouter();
+        long t1 = System.currentTimeMillis();
+        System.out.println("ContentPackageExtractionUtil#directoryUploadAsync | sending for upload extracted package: " + t1);
+		router.tell(request, router);
+        long t2 = System.currentTimeMillis();
+        System.out.println("ContentPackageExtractionUtil#directoryUploadAsync | returning for upload extracted package: " + t2 + ", Time Difference: " + (t1-t2));
 	}
 
 	/**
